+++ src/XRoadLib/Schema/EmptyContentDefinition.cs
@@ -1,0 +1,9 @@
+﻿namespace XRoadLib.Schema
+{
+    public class EmptyContentDefinition : ContentDefinition
+    {
+        public EmptyContentDefinition(ParticleDefinition particle, string runtimeName)
+            : base(particle)
+        {}
+    }
+}