+++ src/XRoadLib/ServiceExecutionOptions.cs
@@ -8,6 +8,8 @@ namespace XRoadLib
     {
         public string RequestNamespace { get; set; }
         public IServiceMap ServiceMap { get; set; }
+        public string OperationName { get; set; }
+        public uint? Version { get; set; }
         public EventHandler<XRoadRequestEventArgs> BeforeRequest;
         public EventHandler<XRoadResponseEventArgs> BeforeDeserialize;
     }